# specs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/orders": {
      "post": {
        "description": "Works for both guest and authenticated users. Provide guestEmail for guest checkout.",
        "operationId": "OrdersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Place a new order",
        "tags": [
          "Orders"
        ]
      }
    },
    "/orders/{orderNumber}": {
      "get": {
        "operationId": "OrdersController_findByOrderNumber",
        "parameters": [
          {
            "name": "orderNumber",
            "required": true,
            "in": "path",
            "schema": {
              "example": "ORD-20240101-0001",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Order not found"
          }
        },
        "summary": "Look up an order by order number",
        "tags": [
          "Orders"
        ]
      }
    },
...
    "/admin/orders/{id}": {
      "get": {
        "operationId": "AdminOrdersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Order UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminOrderResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Admin role required"
          },
          "404": {
            "description": "Order not found"
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Get an order by ID (admin view)",
        "tags": [
          "Admin – Orders"
        ]
      }
    },
...
  },
  "info": {
    "title": "O2Shop API",
    "description": "REST API for the O2Shop e-commerce platform. Public endpoints need no authentication. Customer endpoints require a JWT bearer token (from POST /auth/login). Admin endpoints additionally require the `admin` role.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "access_token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Paste the access_token value from POST /auth/login."
      }
    },
    "schemas": {
...
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "message": {
            "example": [
              "email must be an email"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ]
      },
...
      "AddressResponseDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "Jane"
          },
          "lastName": {
            "type": "string",
            "example": "Doe"
          },
          "company": {
            "type": "string",
            "example": "Acme Corp"
          },
          "address1": {
            "type": "string",
            "example": "123 Main St"
          },
          "address2": {
            "type": "string",
            "example": "Apt 4B"
          },
          "city": {
            "type": "string",
            "example": "New York"
          },
          "country": {
            "type": "string",
            "example": "US"
          },
          "province": {
            "type": "string",
            "example": "NY"
          },
          "postalCode": {
            "type": "string",
            "example": "10001"
          },
          "phone": {
            "type": "string",
            "example": "+1 555 000 0000"
          }
        },
        "required": [
          "firstName",
          "lastName",
          "address1",
          "city",
          "country",
          "province",
          "postalCode",
          "phone"
        ]
      },
...
      "AddressDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "Jane"
          },
          "lastName": {
            "type": "string",
            "example": "Doe"
          },
          "company": {
            "type": "string",
            "example": "Acme Corp"
          },
          "address1": {
            "type": "string",
            "example": "123 Main St"
          },
          "address2": {
            "type": "string",
            "example": "Apt 4B"
          },
          "city": {
            "type": "string",
            "example": "New York"
          },
          "country": {
            "type": "string",
            "example": "US",
            "description": "ISO 3166-1 alpha-2 country code"
          },
          "province": {
            "type": "string",
            "example": "NY"
          },
          "postalCode": {
            "type": "string",
            "example": "10001"
          },
          "phone": {
            "type": "string",
            "example": "+1 555 000 0000"
          }
        },
        "required": [
          "firstName",
          "lastName",
          "address1",
          "city",
          "country",
          "province",
          "postalCode",
          "phone"
        ]
      },
...
      "CreateOrderItemDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string",
            "format": "uuid"
          },
          "variantSku": {
            "type": "string",
            "example": "BLU-M-001"
          },
          "quantity": {
            "type": "number",
            "minimum": 1,
            "example": 2
          }
        },
        "required": [
          "productId",
          "variantSku",
          "quantity"
        ]
      },
      "CreateOrderDto": {
        "type": "object",
        "properties": {
          "guestEmail": {
            "type": "string",
            "format": "email",
            "example": "guest@example.com",
            "description": "Required for guest checkout (no logged-in account)"
          },
          "guestFirstName": {
            "type": "string",
            "example": "Jane"
          },
          "guestLastName": {
            "type": "string",
            "example": "Doe"
          },
          "shippingMethodId": {
            "type": "string",
            "format": "uuid",
            "description": "ID from GET /shipping-methods"
          },
          "shippingAddress": {
            "$ref": "#/components/schemas/AddressDto"
          },
          "billingIsSameAsShipping": {
            "type": "boolean",
            "default": false,
            "example": false
          },
          "billingAddress": {
            "$ref": "#/components/schemas/AddressDto"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateOrderItemDto"
            }
          }
        },
        "required": [
          "shippingMethodId",
          "shippingAddress",
          "billingAddress",
          "items"
        ]
      },
      "PaymentStatus": {
        "type": "string",
        "enum": [
          "pending",
          "paid",
          "failed",
          "refunded"
        ]
      },
      "FulfillmentStatus": {
        "type": "string",
        "enum": [
          "unfulfilled",
          "fulfilled",
          "partially_fulfilled",
          "cancelled"
        ]
      },
      "OrderItemResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "productId": {
            "type": "string",
            "format": "uuid"
          },
          "productName": {
            "type": "string",
            "example": "Blue Widget"
          },
          "productSku": {
            "type": "string",
            "example": "BLU-M-001"
          },
          "productPrice": {
            "type": "number",
            "example": 29.99
          },
          "productCurrency": {
            "type": "string",
            "example": "USD"
          },
          "quantity": {
            "type": "number",
            "example": 2
          },
          "total": {
            "type": "number",
            "example": 59.98
          }
        },
        "required": [
          "id",
          "productName",
          "productSku",
          "productPrice",
          "productCurrency",
          "quantity",
          "total"
        ]
      },
      "OrderResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "orderNumber": {
            "type": "string",
            "example": "ORD-20240101-0001"
          },
          "paymentStatus": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PaymentStatus"
              }
            ]
          },
          "fulfillmentStatus": {
            "allOf": [
              {
                "$ref": "#/components/schemas/FulfillmentStatus"
              }
            ]
          },
          "totalAmount": {
            "type": "number",
            "example": 65.97
          },
          "totalCurrency": {
            "type": "string",
            "example": "USD"
          },
          "shippingMethodName": {
            "type": "string",
            "example": "Standard Shipping"
          },
          "shippingPrice": {
            "type": "number",
            "example": 5.99
          },
          "shippingCurrency": {
            "type": "string",
            "example": "USD"
          },
          "shippingAddress": {
            "$ref": "#/components/schemas/AddressResponseDto"
          },
          "billingAddress": {
            "$ref": "#/components/schemas/AddressResponseDto"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItemResponseDto"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "orderNumber",
          "paymentStatus",
          "fulfillmentStatus",
          "totalAmount",
          "totalCurrency",
          "shippingMethodName",
          "shippingPrice",
          "shippingCurrency",
          "shippingAddress",
          "billingAddress",
          "items",
          "createdAt"
        ]
      },
      "AdminOrderResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "orderNumber": {
            "type": "string",
            "example": "ORD-20240101-0001"
          },
          "paymentStatus": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PaymentStatus"
              }
            ]
          },
          "fulfillmentStatus": {
            "allOf": [
              {
                "$ref": "#/components/schemas/FulfillmentStatus"
              }
            ]
          },
          "totalAmount": {
            "type": "number",
            "example": 65.97
          },
          "totalCurrency": {
            "type": "string",
            "example": "USD"
          },
          "shippingMethodName": {
            "type": "string",
            "example": "Standard Shipping"
          },
          "shippingPrice": {
            "type": "number",
            "example": 5.99
          },
          "shippingCurrency": {
            "type": "string",
            "example": "USD"
          },
          "shippingAddress": {
            "$ref": "#/components/schemas/AddressResponseDto"
          },
          "billingAddress": {
            "$ref": "#/components/schemas/AddressResponseDto"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItemResponseDto"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "guestEmail": {
            "type": "string",
            "example": "guest@example.com"
          },
          "guestFirstName": {
            "type": "string",
            "example": "Jane"
          },
          "guestLastName": {
            "type": "string",
            "example": "Doe"
          },
          "paymentProviderId": {
            "type": "string"
          },
          "paymentProviderRef": {
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "orderNumber",
          "paymentStatus",
          "fulfillmentStatus",
          "totalAmount",
          "totalCurrency",
          "shippingMethodName",
          "shippingPrice",
          "shippingCurrency",
          "shippingAddress",
          "billingAddress",
          "items",
          "createdAt",
          "updatedAt"
        ]
      },
...
    }
  }
}